+++ src/tokenServer/JsonImageCache.cs
@@ -1,16 +1,17 @@
-﻿using System;
+﻿using GaRyan2.Utilities;
+using Newtonsoft.Json;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using Newtonsoft.Json;
 
 namespace tokenServer
 {
     public static class JsonImageCache
     {
-        public static Dictionary<string, CacheImage> ImageCache;
+        public static Dictionary<string, CacheImage> ImageCache = new Dictionary<string, CacheImage>();
         private static readonly object _cacheLock = new object();
-        public static bool cacheImages;
+        public static bool cacheImages => cacheRetention > 0;
         public static int cacheRetention;
 
         static JsonImageCache()
@@ -18,26 +19,23 @@ namespace tokenServer
             Load();
         }
 
-        private static void Load()
+        public static void Load()
         {
             if (File.Exists(Helper.Epg123ImageCachePath))
             {
-                using (var reader = File.OpenText(Helper.Epg123ImageCachePath))
-                {
-                    var serializer = new JsonSerializer();
-                    ImageCache = (Dictionary<string, CacheImage>)serializer.Deserialize(reader, typeof(Dictionary<string, CacheImage>));
-                }
+                Helper.ReadJsonFile(Helper.Epg123ImageCachePath, typeof(Dictionary<string, CacheImage>));
             }
             else ImageCache = new Dictionary<string, CacheImage>();
         }
 
         public static void Cleanup()
         {
+            if (!cacheImages) return;
             var markedForDelete = ImageCache.Where(arg => arg.Value.LastUsed + TimeSpan.FromDays(cacheRetention) < DateTime.Now)
                 .Select(arg => arg.Key).ToList();
             foreach (var image in markedForDelete)
             {
-                var path = $"{Helper.Epg123ImageCache}\\{image.Substring(0, 1)}\\{image}";
+                var path = $"{Helper.Epg123ImageCache}{image.Substring(0, 1)}\\{image}";
                 try
                 {
                     if (File.Exists(path)) File.Delete(path);
@@ -52,11 +50,7 @@ namespace tokenServer
             lock (_cacheLock)
             {
                 Cleanup();
-                using (var writer = File.CreateText(Helper.Epg123ImageCachePath))
-                {
-                    var serializer = new JsonSerializer { NullValueHandling = NullValueHandling.Ignore, Formatting = Formatting.None };
-                    serializer.Serialize(writer, ImageCache);
-                }
+                Helper.WriteJsonFile(ImageCache, Helper.Epg123ImageCachePath);
             }
         }
 
@@ -64,7 +58,7 @@ namespace tokenServer
         {
             // remove "/image/" from beginning
             filename = filename.Substring(7);
-            var location = $"{Helper.Epg123ImageCache}\\{filename.Substring(0, 1)}\\{filename}";
+            var location = $"{Helper.Epg123ImageCache}{filename.Substring(0, 1)}\\{filename}";
             lock (_cacheLock)
             {
                 var fi = new FileInfo(location);
@@ -72,7 +66,7 @@ namespace tokenServer
                 {
                     if (fi.Exists)
                     {
-                        ImageCache[filename].LastUsed = DateTime.Now;
+                        if (DateTime.Now - ImageCache[filename].LastUsed > TimeSpan.FromHours(24)) ImageCache[filename].LastUsed = DateTime.Now;
                         if (ImageCache[filename].ByteSize == 0) ImageCache[filename].ByteSize = fi.Length;
                         return fi;
                     }
@@ -85,25 +79,25 @@ namespace tokenServer
             }
         }
 
-        public static void AddImageToCache(string filename, DateTime lastModified, long size)
+        public static void AddImageToCache(string filename, DateTimeOffset lastModified, long size)
         {
             lock (_cacheLock)
             {
                 if (ImageCache.ContainsKey(filename))
                 {
                     ImageCache[filename].LastUsed = DateTime.Now;
-                    ImageCache[filename].LastModified = lastModified;
+                    ImageCache[filename].LastModified = lastModified.LocalDateTime;
                     ImageCache[filename].ByteSize = size;
                 }
-                else ImageCache.Add(filename, new CacheImage { LastUsed = DateTime.Now, LastModified = lastModified, ByteSize = size});
+                else ImageCache.Add(filename, new CacheImage { LastUsed = DateTime.Now, LastModified = lastModified.LocalDateTime, ByteSize = size});
             }
         }
 
         public static void GetAllImageSizes()
         {
-            foreach (var image in ImageCache.Where(arg => arg.Value.ByteSize == 0))
+            foreach (var image in ImageCache.Where(arg => arg.Value.ByteSize == 0).ToList())
             {
-                var fi = new FileInfo($"{Helper.Epg123ImageCache}\\{image.Key.Substring(0, 1)}\\{image.Key}");
+                var fi = new FileInfo($"{Helper.Epg123ImageCache}{image.Key.Substring(0, 1)}\\{image.Key}");
                 if (fi.Exists)
                 {
                     image.Value.ByteSize = fi.Length;
@@ -114,7 +108,7 @@ namespace tokenServer
         public static void AddImagesMissingInCacheFile()
         {
             var files = Directory.GetFiles($"{Helper.Epg123ImageCache}", "*.*", SearchOption.AllDirectories);
-            foreach (var file in files.Where(arg => arg.EndsWith("jpg") || arg.EndsWith("png")))
+            foreach (var file in files.Where(arg => arg.EndsWith("jpg") || arg.EndsWith("png")).ToList())
             {
                 var fi = new FileInfo(file);
                 if (ImageCache.ContainsKey(fi.Name)) continue;
@@ -122,21 +116,21 @@ namespace tokenServer
             }
         }
 
-        public static bool IsImageRecent(string filename, DateTime ifModifiedSince)
+        public static bool IsImageRecent(string filename, DateTimeOffset ifModifiedSince)
         {
             lock (_cacheLock)
             {
                 if (!ImageCache.ContainsKey(filename)) return false;
                 if (ImageCache[filename].LastModified == DateTime.MinValue)
                 {
-                    var info = new FileInfo($"{Helper.Epg123ImageCache}\\{filename.Substring(0, 1)}\\{filename}");
-                    ImageCache[filename].LastModified = info.LastWriteTimeUtc;
+                    var info = new FileInfo($"{Helper.Epg123ImageCache}{filename.Substring(0, 1)}\\{filename}");
+                    ImageCache[filename].LastModified = info.LastWriteTime;
                 }
-                if (ifModifiedSince == DateTime.MinValue && DateTime.UtcNow - ImageCache[filename].LastModified < TimeSpan.FromDays(30))
+                if (ifModifiedSince == DateTime.MinValue && DateTime.Now - ImageCache[filename].LastModified < TimeSpan.FromDays(30))
                 {
                     return true;
                 }
-                if (ImageCache[filename].LastUsed.ToLocalTime() + TimeSpan.FromHours(24) > DateTime.Now) return true;
+                if (ImageCache[filename].LastUsed + TimeSpan.FromHours(24) > DateTime.Now) return true;
             }
             return false;
         }
